Changed: move header in shared folder
