--- src/app/components/header/header.component.html
@@ -1,37 +1,0 @@
-<nav class="navbar navbar-default">
-    <div class="container">
-
-        <div class="navbar-header">
-            <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1" aria-expanded="false">
-                <span class="sr-only">Toggle navigation</span>
-                <span class="icon-bar"></span>
-                <span class="icon-bar"></span>
-                <span class="icon-bar"></span>
-            </button>
-            <h1 class="no-margin">
-                <a class="navbar-brand" routerLink="/dashboard">{{ title }}</a>
-            </h1>
-        </div>
-
-        <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
-            <ul class="nav navbar-nav">
-                <li
-                    *ngIf="authService.getLoggedUser() != null &&
-                    authService.getLoggedUser().isAdmin()">
-                    <a routerLink="/admin" routerLinkActive="active">Admin</a>
-                </li>
-                <li *ngIf="!authService.isAuthenticated()">
-                    <a routerLink="/login" routerLinkActive="active">Login</a>
-                </li>
-                <li *ngIf="!authService.isAuthenticated()">
-                    <a routerLink="/register" routerLinkActive="active">Register</a>
-                </li>
-            </ul>
-            <ul class="nav navbar-nav right">
-                <li *ngIf="authService.isAuthenticated()">
-                    <a href="javascript:;" (click)="logout()">Logout</a>
-                </li>
-            </ul>
-        </div>
-    </div>
-</nav>
