+++ src/app/shared/header.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 
-import { AuthService } from '../../core/auth.service';
+import { AuthService } from '../core/auth.service';
 
 @Component({
     selector: 'xp-header',
